requirements.txt
rdflib==6.2.0
fastapi==0.85.0
uvicorn==0.18.3
python_dotenv==0.21.0